Transform root symbols and don't transform properties
When taking away the AMD closure, symbols that were private
would become global, which would conflict with symbols that
become global in other AMD modules. To fix this, we namespace
all symbols that become global.

Properties like obj.property and {property: value} will not
be transformed anymore with this commit.